Add files via upload
This is not a large sample, but consider that these are college students asked to take the time to fill out a survey, with comments. Doing so will not benefit them in the slightest. There is reason to believe that this teaching approach works. Students come away feeling as if they've learned something valuable, useful, and permanent. It's a win-win.